# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TheDotFactoryChanges - OpenAPI 3.0
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: https://localhost:{port}/api/{basePath}
    variables:
      port:
        enum:
          - '59572'
        default: '59572'
      basePath:
        enum:
          - 'v1'
        default: v1
paths:
...
  /configurationNames:
    get:
      tags:
        - configuration
      summary: Get available configuration names
      description: Get available configuration names
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigurationName'
        '401':
          description: User is not authorized
        '500':
          description: Server error
  /configurations:
    get:
      tags:
        - configuration
      summary: Get available configurations
      description: Get available configurations
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Configuration'
        '401':
          description: User is not authorized
        '500':
          description: Server error
    post:
      tags:
        - configuration
      summary: Create configurations
      description: Create configurations
      requestBody:
        description: Configuration
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Configuration'
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request
        '401':
          description: User not authorized
        '500':
          description: Server error
  /configurations/{configurationId}:
    get:
      tags:
        - configuration
      summary: Get configuration by id
      description: Get configuration by id
      parameters:
        - name: configurationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Configuration'
        '401':
          description: User not authorized
        '404':
          description: Configuration not found
        '500':
          description: Server error
    put:
      tags:
        - configuration
      summary: Update configuration
      description: Update configuration
      parameters:
        - name: configurationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Configuration
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Configuration'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: User not authorized
        '404':
          description: Configuration not found
        '500':
          description: Server error
    delete:
      tags:
        - configuration
      summary: Delete configuration by id
      description: Delete configuration by id
      parameters:
        - name: configurationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: User is not authorized
        '404':
          description: Configuration not found
        '500':
          description: Server error
...
components:
  schemas:
...
    ConfigurationName:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: string
    Configuration:
      required:
        - id
        - userId
        - commentVariableName
        - commentCharVisualizer
        - commentCharDescriptor
        - commentStyle
        - bmpVisualizerChar
        - rotation
        - flipHorizontal
        - flipVertical
        - paddingRemovalHorizontal
        - paddingRemovalVertical
        - bitLayout
        - byteOrder
        - byteFormat
        - generateLookupArray
        - generateSpaceCharacterBitmap
        - spaceGenerationPixels
        - minHeight
        - varNfBitmaps
        - varNfCharInfo
        - varNfFontInfo
        - varNfWidth
        - varNfHeight
        - displayName
        - commentStartString
        - commentBlockEndString
        - commentBlockMiddleString
        - commentEndString
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        userId:
          type: integer
          format: int64
          example: 10
        commentVariableName:
          type: boolean
          example: true
        commentCharVisualizer:
          type: boolean
          example: true
        commentCharDescriptor:
          type: boolean
          example: true
        commentStyle:
          type: string
          enum:
            - c
            - cpp
        bmpVisualizerChar:
          type: string
          example: string
        rotation:
          type: string
          enum:
            - RotateZero
            - RotateNinety
            - RotateOneEighty
            - RotateTwoSeventy
        flipHorizontal:
          type: boolean
          example: true
        flipVertical:
          type: boolean
          example: true
        paddingRemovalHorizontal:
          type: string
          enum:
            - None
            - Tighest
            - Fixed
        paddingRemovalVertical:
          type: string
          enum:
            - None
            - Tighest
            - Fixed
        bitLayout:
          type: string
          enum:
            - RowMajor
            - ColumnMajor
        byteOrder:
          type: string
          enum:
            - LsbFirst
            - MsbFirst
        byteFormat:
          type: string
          enum:
            - Binary
            - Hex
        generateLookupArray:
          type: boolean
          example: true
        generateSpaceCharacterBitmap:
          type: boolean
          example: true
        spaceGenerationPixels:
          type: integer
          format: int64
          example: 10
        minHeight:
          type: integer
          format: int64
          example: 10
        varNfBitmaps:
          type: string
          example: string
        varNfCharInfo:
          type: string
          example: string
        varNfFontInfo:
          type: string
          example: string
        varNfWidth:
          type: string
          example: string
        varNfHeight:
          type: string
          example: string
        displayName:
          type: string
          example: string
        commentStartString:
          type: string
          example: string
        commentBlockEndString:
          type: string
          example: string
        commentBlockMiddleString:
          type: string
          example: string
        commentEndString:
          type: string
          example: string
...